scrub cert commands and move to point of use
