+++ pkg/oc/cli/admin/cert/encrypt.go
@@ -18,8 +18,6 @@ import (
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	kcmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 	"k8s.io/kubernetes/pkg/kubectl/util/templates"
-
-	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
 )
 
 const EncryptCommandName = "encrypt"
@@ -152,10 +150,10 @@ func (o *EncryptOptions) Encrypt() error {
 	var key []byte
 	switch {
 	case len(o.KeyFile) > 0:
-		if block, ok, err := BlockFromFile(o.KeyFile, configapi.StringSourceKeyBlockType); err != nil {
+		if block, ok, err := BlockFromFile(o.KeyFile, StringSourceKeyBlockType); err != nil {
 			return err
 		} else if !ok {
-			return fmt.Errorf("%s does not contain a valid PEM block of type %q", o.KeyFile, configapi.StringSourceKeyBlockType)
+			return fmt.Errorf("%s does not contain a valid PEM block of type %q", o.KeyFile, StringSourceKeyBlockType)
 		} else if len(block.Bytes) == 0 {
 			return fmt.Errorf("%s does not contain a key", o.KeyFile)
 		} else {
@@ -172,7 +170,7 @@ func (o *EncryptOptions) Encrypt() error {
 	}
 
 	// Encrypt
-	dataBlock, err := x509.EncryptPEMBlock(rand.Reader, configapi.StringSourceEncryptedBlockType, data, key, x509.PEMCipherAES256)
+	dataBlock, err := x509.EncryptPEMBlock(rand.Reader, StringSourceEncryptedBlockType, data, key, x509.PEMCipherAES256)
 	if err != nil {
 		return err
 	}
@@ -198,7 +196,7 @@ func (o *EncryptOptions) Encrypt() error {
 
 	// Write key
 	if len(o.GenKeyFile) > 0 {
-		keyBlock := &pem.Block{Bytes: key, Type: configapi.StringSourceKeyBlockType}
+		keyBlock := &pem.Block{Bytes: key, Type: StringSourceKeyBlockType}
 		if err := BlockToFile(o.GenKeyFile, keyBlock, os.FileMode(0600)); err != nil {
 			return err
 		}
