+++ test/util/server/deprecated_openshift/deprecatedcerts/signer_cert_args.go
@@ -1,4 +1,4 @@
-package signercertoptions
+package deprecatedcerts
 
 import (
 	"fmt"
